Added Collector Functionality
Filled in the Collector class and made commands for the collector. Not
sure if the gearbox has to be accounted for in the code, so right now it
is running as if it were just a motor. Also set the PID source parameter
for the encoders in the Drivetrain class to Distance. Added a bunch of
static variables to the RobotMap class.