Added command groups, buttons, and improved PID
Added a retract and shoot command group for catapault, an extend and
drive forward group for the collector, added a bunch of buttons in OI,
and edited the PID controllers for the brakes in drivetrain so that
there are only 2 PID controllers, one for the left side, one for the
right. Also added setpoints to the controllers when the brakes are
enabled.